Reorganized test artifacts
